Improved Subscription / Resource models
Updated the location and resource objects to be a bit more generic as well as
renamed some files in the subscriber models to be a bit more clear while also
adding some class-file comments. Also added some message and conformation
classes for de-registration on th provider and subscriber.

Updated .gitignore with vim and IntelliJ files